requirements.txt
absl-py==0.7.1
aiohttp==3.7.4.post0
altgraph==0.16.1
asgiref==3.2.7
astor==0.7.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Keras==2.3.1
Keras-Applications==1.0.8
keras-contrib==2.0.8
Keras-Preprocessing==1.1.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.1.1
nilearn==0.5.2
numpy==1.17.0
opencv-contrib-python==4.5.1.48
opencv-python==3.4.4.19
opt-einsum==3.2.0
pandas==0.25.0
Pillow==5.3.0
pydicom==1.3.0
scikit-image==0.15.0
scikit-learn==0.22
scikit-plot==0.3.7
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
shap==0.39.0
SimpleITK==1.2.0
simplejson==3.17.2
sip==4.19.8
six==1.11.0
sklearn==0.0
slicer==0.0.7
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
urllib3==1.25.6
vtk==9.0.3
